backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
opencv-python==4.9.0.80
mediapipe==0.10.9
numpy==1.24.3
pillow==10.2.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
cryptography==42.0.0
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.1.2
scikit-image==0.22.0
scipy==1.11.4
pydantic>=2.7.0
pydantic-settings>=2.1.0
passlib==1.7.4
fastapi-mail==1.4.1
slowapi==0.1.9
redis==5.0.1
httpx==0.26.0
pytest==8.0.0
pytest-asyncio==0.23.5
psutil==5.9.8
reportlab>=4.0.0
